hot fix for throwbackloader and dotnet
Hot patch
updated the Throwbackloader link (old got removed and it also got updated).  
Changed how the downloader sees dotnet (checks if you have dotnet 9 or higher if not it goes to the dotnet download page to download).  
Next update V7 will be a complete rewrite of the bat file with better optimization and auto-updater.